Unnecessary entities deleted, Id from Pet entiti from Int32 to Int64

+++ Controllers/PetsController.cs
@@ -60,7 +60,7 @@ namespace GURU_CoMM.Controllers
         }
 
         [HttpGet("{id:int}", Name = "GetPet")]
-        public async Task<ActionResult<PetDto>> Get(int id)
+        public async Task<ActionResult<PetDto>> Get(long id)
         {
             var entidad = await context.Pets.FirstOrDefaultAsync(x => x.Id == id);
